Progress
Provides schemas for Posts and Stores, alongside adding posts to the User schema.

Now all schemas are resolved by default, which is more intuitive, although bad for performance. For our case of a demonstration, testing app, where we should have at most hundreds of posts and users, it shouldn't be an issue.