19 - Backend Schemas (#23)
* Progress

Provides schemas for Posts and Stores, alongside adding posts to the User schema.

Now all schemas are resolved by default, which is more intuitive, although bad for performance. For our case of a demonstration, testing app, where we should have at most hundreds of posts and users, it shouldn't be an issue.

* Solve dependency issue

Storing posts in the user is pointless as you can simply query all posts and filter by the author, anyway.

* Group into typescript/mongoose schemas + labels

* Add .draw.io UML Diagram (.xml)

* Progress

Provides schemas for Posts and Stores, alongside adding posts to the User schema.

Now all schemas are resolved by default, which is more intuitive, although bad for performance. For our case of a demonstration, testing app, where we should have at most hundreds of posts and users, it shouldn't be an issue.

* Solve dependency issue

Storing posts in the user is pointless as you can simply query all posts and filter by the author, anyway.

* Group into typescript/mongoose schemas + labels

---------

Co-authored-by: KrakenM (T) <moemyatag05@gmail.com>